middleware: Extract `apply_axum_middleware()` function
